refactor(metrics): migrate specs to registry (#472)
Physically isolate 6 pipeline primitives under _primitives/ and refactor
family modules to use the class-based @metric decorator.

Remove legacy __metric_specs__ tuples, privatize pipeline functions, and
adjust tests to import from _primitives. Fix MetricBase.__call__ typing
to support non-MetricResult returns from primitives.

Refs #472
