+++ docs/api/metrics/caar.md
@@ -6,7 +6,6 @@ title: factrix.metrics.caar
     options:
       show_root_members_full_path: true
       members:
-        - compute_caar
         - caar
         - bmp_test
